requirements.txt
firebase-admin
google-generativeai
python-dotenv